backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
PyJWT
orjson
psycopg2-binary
python-dotenv
passlib[bcrypt]
pydantic-settings
python-multipart
PyMuPDF
aiosqlite
requests
gunicorn
httpx
rich
websockets